Modifying docstrings and extracting time step calculation function, following suggestions received.
