+++ DSSE/environment/generator/dynamic_probability.py
@@ -1,6 +1,6 @@
 import numpy as np
-from numpy.linalg import norm
 from numba import njit
+from DSSE.environment.time_step import calculate_time_step
 
 
 class ProbabilityMatrix:
@@ -150,43 +150,7 @@ class ProbabilityMatrix:
         return probabilities
 
     def update_time_step_relation(self, time_step: float, cell_size: float) -> None:
-        self.time_step_relation = self.calculate_time_step(time_step, self.movement_vector, cell_size)
-
-    def calculate_time_step(
-            self,
-            time_step: float,
-            speed: tuple[float],
-            cell_size: float
-        ) -> float:
-        """
-        Parameters:
-        ----------
-        time_step: float
-            Time step in seconds
-        person_speed: tuple[float]
-            Speed of the person in the water in m/s (x and y components)
-        cell_size: float
-            Size of the cells in meters
-
-        Returns:
-        -------
-        int
-            Time step realtion in number of iterations
-        """
-        speed_magnitude = self.calculate_vector_magnitude(speed)
-        return cell_size / speed_magnitude / time_step
-
-    def calculate_vector_magnitude(self, vector: tuple[float]) -> float:
-        """
-        Args:
-        vector: tuple[float]
-            Vector with x and y components
-
-        Returns:
-        magnitude : float
-            Magnitude is in m/s
-        """
-        return np.linalg.norm(vector)
+        self.time_step_relation = calculate_time_step(time_step, self.movement_vector, cell_size)
 
     def get_matrix(self):
         return self.map_prob
